requirements.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
datasets
ftfy
tensorboard
Jinja2
peft==0.7.0
diffusers
wandb